filtragem de material na página principal, tabela dos itens que foram emprestados, correção no botão de devolver

+++ app/Http/Controllers/EmprestimoController.php
@@ -42,23 +42,27 @@ class EmprestimoController extends Controller
     public function devolver(Request $request, Emprestimo $emprestimo)
     {
         $ids = $request->itens; //capturando os ids  dos itens que foram passados pelo usuário através da checkbox
-
-        if (sizeof($ids) == sizeof($emprestimo->itens)) {
-            //esta comparando se a quantidade de itens, se for a mesma quantidade significa que todos os itens do empréstimo foram devolvido,logo eu dissocio apenas os itens e apago o empréstimo
-            foreach ($emprestimo->itens as $item) {
-                $item->disponibilidade = true;
-                $item->save();
-            }
-            $emprestimo->itens()->detach();
-            $emprestimo->delete();
-        } else {
-            //se a quantidade não for igual, então nem todos os itens foram devolvidos, logo dissocio apenas os itens que foram devolvidos
-            for ($i = 0; $i < sizeof($ids); $i++) {
-                $emprestimo->itens()->detach($ids[$i]);
-                foreach (Item::find($ids)as $item) {
-                    $item->disponibilidade = true; $item->save();
+        if ($ids) {
+            if (sizeof($ids) == sizeof($emprestimo->itens)) {
+                //esta comparando se a quantidade de itens, se for a mesma quantidade significa que todos os itens do empréstimo foram devolvido,logo eu dissocio apenas os itens e apago o empréstimo
+                foreach ($emprestimo->itens as $item) {
+                    $item->disponibilidade = true;
+                    $item->save();
+                }
+                $emprestimo->itens()->detach();
+                $emprestimo->delete();
+            } else {
+                //se a quantidade não for igual, então nem todos os itens foram devolvidos, logo dissocio apenas os itens que foram devolvidos
+                for ($i = 0; $i < sizeof($ids); $i++) {
+                    $emprestimo->itens()->detach($ids[$i]);
+                    foreach (Item::find($ids) as $item) {
+                        $item->disponibilidade = true;
+                        $item->save();
+                    }
                 }
             }
+        }else{
+            return back();
         }
 
         return redirect(route('emprestimos.index'));
